JENA-1454: Introduce builder pattern for result set reading and writing.
