+++ jena-arq/src/main/java/org/apache/jena/riot/resultset/ResultSetWriterRegistry.java
@@ -18,12 +18,7 @@
 
 package org.apache.jena.riot.resultset;
 
-import static org.apache.jena.riot.resultset.ResultSetLang.SPARQLResultSetCSV ;
-import static org.apache.jena.riot.resultset.ResultSetLang.SPARQLResultSetJSON ;
-import static org.apache.jena.riot.resultset.ResultSetLang.SPARQLResultSetTSV ;
-import static org.apache.jena.riot.resultset.ResultSetLang.SPARQLResultSetText ;
-import static org.apache.jena.riot.resultset.ResultSetLang.SPARQLResultSetThrift ;
-import static org.apache.jena.riot.resultset.ResultSetLang.SPARQLResultSetXML ;
+import static org.apache.jena.riot.resultset.ResultSetLang.*;
 
 import java.io.OutputStream ;
 import java.io.Writer ;
@@ -35,9 +30,13 @@ import org.apache.jena.atlas.lib.NotImplemented ;
 import org.apache.jena.query.ResultSet ;
 import org.apache.jena.riot.Lang ;
 import org.apache.jena.riot.RiotException ;
-import org.apache.jena.riot.thrift.BinRDF ;
+import org.apache.jena.riot.resultset.rw.ResultSetWriterJSON;
+import org.apache.jena.riot.resultset.rw.ResultSetWriterThrift;
+import org.apache.jena.riot.resultset.rw.ResultSetWriterXML;
 import org.apache.jena.sparql.core.Prologue ;
-import org.apache.jena.sparql.resultset.* ;
+import org.apache.jena.sparql.resultset.CSVOutput;
+import org.apache.jena.sparql.resultset.TSVOutput;
+import org.apache.jena.sparql.resultset.TextOutput;
 import org.apache.jena.sparql.serializer.SerializationContext ;
 import org.apache.jena.sparql.util.Context ;
 
@@ -46,9 +45,14 @@ public class ResultSetWriterRegistry {
     private static Map<Lang, ResultSetWriterFactory> registry = new HashMap<>() ;
     
     /** Lookup a {@link Lang} to get the registered {@link ResultSetReaderFactory} (or null) */
-    public static ResultSetWriterFactory lookup(Lang lang) {
+    public static ResultSetWriterFactory getFactory(Lang lang) {
         Objects.requireNonNull(lang) ;
         return registry.get(lang) ;
+    }
+
+    public static boolean isRegistered(Lang lang) {
+        Objects.requireNonNull(lang) ;
+        return registry.containsKey(lang) ;
     }
 
     /** Register a {@link ResultSetReaderFactory} for a {@link Lang} */
@@ -65,45 +69,21 @@ public class ResultSetWriterRegistry {
         initialized = true ;
 
         ResultSetWriterFactory factory = new ResultSetWriterFactoryStd() ;
-        register(SPARQLResultSetXML,    factory) ;
-        register(SPARQLResultSetJSON,   factory) ;
+        register(SPARQLResultSetXML,    ResultSetWriterXML.factory) ;
+        register(SPARQLResultSetJSON,   ResultSetWriterJSON.factory) ;
+        register(SPARQLResultSetThrift, ResultSetWriterThrift.factory) ;
         register(SPARQLResultSetCSV,    factory) ;
         register(SPARQLResultSetTSV,    factory) ;
-        register(SPARQLResultSetThrift, new ResultSetWriterThriftFactory()) ;
         register(SPARQLResultSetText,   factory) ;
     }
  
-    private static ResultSetWriter writerXML = new ResultSetWriter() {
-        @Override public void write(OutputStream out, ResultSet resultSet, Context context) { 
-            XMLOutput xOut = new XMLOutput(null) ;
-            xOut.format(out, resultSet) ;
-        }
-        @Override public void write(Writer out, ResultSet resultSet, Context context) {throw new NotImplemented("Writer") ; }
-        @Override public void write(OutputStream out, boolean result, Context context) {
-            XMLOutput xOut = new XMLOutput(null);
-            xOut.format(out, result);
-        }
-    } ;
-
-    private static ResultSetWriter writerJSON = new ResultSetWriter() {
-        @Override public void write(OutputStream out, ResultSet resultSet, Context context) {
-            JSONOutput jOut = new JSONOutput() ;
-            jOut.format(out, resultSet) ; 
-        }
-        @Override public void write(Writer out, ResultSet resultSet, Context context) {throw new NotImplemented("Writer") ; }
-        @Override public void write(OutputStream out, boolean result, Context context) {
-            JSONOutput jOut = new JSONOutput() ;
-            jOut.format(out, result) ; 
-        }
-    } ;
-    
     private static ResultSetWriter writerCSV = new ResultSetWriter() {
         @Override public void write(OutputStream out, ResultSet resultSet, Context context) {
             CSVOutput fmt = new CSVOutput() ;
             fmt.format(out, resultSet) ;
         }
-        @Override public void write(Writer out, ResultSet resultSet, Context context) {throw new NotImplemented("Writer") ; }
-        @Override public void write(OutputStream out, boolean result, Context context) {
+        @Override public void write(Writer out, ResultSet resultSet, Context context)   { throw new NotImplemented("Writer") ; }
+        @Override public void write(OutputStream out, boolean result, Context context)  {
             CSVOutput fmt = new CSVOutput() ;
             fmt.format(out, result) ;
         }
@@ -114,8 +94,8 @@ public class ResultSetWriterRegistry {
             TSVOutput fmt = new TSVOutput() ;
             fmt.format(out, resultSet) ;
         }
-        @Override public void write(Writer out, ResultSet resultSet, Context context) {throw new NotImplemented("Writer") ; }
-        @Override public void write(OutputStream out, boolean result, Context context) {
+        @Override public void write(Writer out, ResultSet resultSet, Context context)   {throw new NotImplemented("Writer") ; }
+        @Override public void write(OutputStream out, boolean result, Context context)  {
             TSVOutput fmt = new TSVOutput() ;
             fmt.format(out, result) ;
         }
@@ -124,7 +104,7 @@ public class ResultSetWriterRegistry {
     private static ResultSetWriter writerNo = new ResultSetWriter() {
         @Override public void write(OutputStream out, ResultSet resultSet, Context context) {}
         @Override public void write(Writer out, ResultSet resultSet, Context context)       {}
-        @Override public void write(OutputStream out, boolean result, Context context) {}
+        @Override public void write(OutputStream out, boolean result, Context context)      {}
     } ;
 
     private static ResultSetWriter writerText = new ResultSetWriter() {
@@ -145,31 +125,12 @@ public class ResultSetWriterRegistry {
         @Override
         public ResultSetWriter create(Lang lang) {
             lang = Objects.requireNonNull(lang, "Language must not be null") ;
-            if ( lang.equals(SPARQLResultSetXML) )      return writerXML ;
-            if ( lang.equals(SPARQLResultSetJSON) )     return writerJSON ;
+//            if ( lang.equals(SPARQLResultSetXML) )      return writerXML ;
+//            if ( lang.equals(SPARQLResultSetJSON) )     return writerJSON ;
             if ( lang.equals(SPARQLResultSetCSV) )      return writerCSV ;
             if ( lang.equals(SPARQLResultSetTSV) )      return writerTSV ;
             if ( lang.equals(SPARQLResultSetText) )     return writerText ;
             throw new RiotException("Lang not registered (ResultSet writer)") ;
-        }
-    }
-    
-    private static class ResultSetWriterThriftFactory implements ResultSetWriterFactory {
-        @Override
-        public ResultSetWriter create(Lang lang) {
-            return new ResultSetWriter() {
-                @Override
-                public void write(OutputStream out, ResultSet resultSet, Context context)
-                { BinRDF.writeResultSet(out, resultSet) ; }
-                
-                @Override
-                public void write(Writer out, ResultSet resultSet, Context context) {
-                    throw new NotImplemented("Writing binary data to a java.io.Writer is not possible") ;
-                }
-                @Override
-                public void write(OutputStream out, boolean result, Context context)
-                { throw new NotImplemented("No Thrift RDF encoding defined for boolean results"); }
-            } ;
         }
     }
 }
