+++ jena-arq/src/main/java/org/apache/jena/sparql/resultset/SPARQLResult.java
@@ -18,13 +18,9 @@
 
 package org.apache.jena.sparql.resultset;
 
-import org.apache.jena.atlas.logging.Log;
-import org.apache.jena.graph.Node;
 import org.apache.jena.query.Dataset;
 import org.apache.jena.query.ResultSet;
 import org.apache.jena.rdf.model.Model;
-import org.apache.jena.sparql.core.Var;
-import org.apache.jena.sparql.engine.binding.BindingMap;
 
 /**
  * The class "ResultSet" is reserved for the SELECT result format. This class
@@ -95,7 +91,7 @@ public class SPARQLResult {
         return resultSet;
     }
 
-    public boolean getBooleanResult() {
+    public Boolean getBooleanResult() {
         if ( !hasBeenSet )
             throw new ResultSetException("Not set");
         if ( !isBoolean() )
@@ -146,18 +142,4 @@ public class SPARQLResult {
         booleanResult = r;
         hasBeenSet = true;
     }
-
-    static protected void addBinding(BindingMap binding, Var var, Node value) {
-        Node n = binding.get(var);
-        if ( n != null ) {
-            // Same - silently skip.
-            if ( n.equals(value) )
-                return;
-            Log.warn(SPARQLResult.class,
-                     String.format("Multiple occurences of a binding for variable '%s' with different values - ignored", var.getName()));
-            return;
-        }
-        binding.add(var, value);
-    }
-
 }
