+++ jena-arq/src/main/java/org/apache/jena/sparql/util/LabelToNodeMap.java
@@ -28,8 +28,7 @@ import org.apache.jena.sparql.ARQConstants ;
 import org.apache.jena.sparql.core.VarAlloc ;
 
 
-/** Map from _:* form to bNodes */
-
+/** Map from _:* form to bNodes or variables. */
 public class LabelToNodeMap
 {
     // Wil be replaced by LabelToNode in RIOT
@@ -44,8 +43,9 @@ public class LabelToNodeMap
     /** Create blank nodes, with the same blank node returned for the same label.  
      * 
      * @return LabelToNodeMap
+     * @deprecated use LabelToNode 
      */
-
+    @Deprecated
     public static LabelToNodeMap createBNodeMap()
     { return new LabelToNodeMap(false, null) ; }
